fix: 대용량 ROFL 파일 업로드 시 변환 성능 개선 및 멈춤 문제 해결
- 전체 파일을 문자열로 변환하는 대신, 필요한 JSON 시작 위치 이후만 UTF-8 변환
- 대형 바이너리 파일 처리 시 발생하던 CPU 부하 및 OOM 문제 예방
- 업로드 성공률 및 서버 안정성 대폭 향상

+++ back/roflparser/src/main/java/com/example/roflparser/domain/Match.java
@@ -1,8 +1,8 @@
 package com.example.roflparser.domain;
 
 import jakarta.persistence.*;
-        import lombok.*;
-        import org.springframework.data.annotation.CreatedDate;
+import lombok.*;
+import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
@@ -28,7 +28,7 @@ public class Match {
     private LocalDateTime gameDatetime;
 
     @Column(name = "game_length")
-    private Integer gameLength;
+    private Long gameLength; // ▶️ Integer -> Long 변경
 
     @CreatedDate
     private LocalDateTime createdAt;
